+++ src/main/java/kr/hhplus/be/server/application/order/repository/OrderRepository.java
@@ -5,6 +5,8 @@ import kr.hhplus.be.server.domain.order.Order;
 import java.util.Optional;
 
 public interface OrderRepository {
+
+
     Order save(Order order);
 
     Optional<Order> findById(long orderId);
