+++ src/main/java/kr/hhplus/be/server/application/product/service/ProductService.java
@@ -1,21 +1,26 @@
 package kr.hhplus.be.server.application.product.service;
 
+import kr.hhplus.be.server.application.order.OrderCriteria;
+import kr.hhplus.be.server.application.product.repository.InventoryRepository;
 import kr.hhplus.be.server.application.product.repository.ProductRepository;
 import kr.hhplus.be.server.application.product.repository.TopSellingProductRepository;
 import kr.hhplus.be.server.domain.order.Order;
+import kr.hhplus.be.server.domain.product.Inventory;
 import kr.hhplus.be.server.domain.product.Product;
 import kr.hhplus.be.server.domain.product.TopSellingProduct;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Map;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class ProductService {
 
     private final ProductRepository productRepository;
+    private final InventoryRepository inventoryRepository;
     private final TopSellingProductRepository topSellingProductRepository;
 
     public Product getById(long id) {
@@ -26,37 +31,50 @@ public class ProductService {
         return productRepository.findAll();
     }
 
+    public List<Order.ProductQuantity> getProductsStock(List<OrderCriteria.OrderProduct> orderProducts) {
+        return orderProducts.stream().map(orderProduct -> {
+            long productId = orderProduct.getProductId();
 
-    public List<Order.ProductQuantity> getProductsStock(Map<Long, Integer> productQuantities) {
-        return productQuantities.entrySet().stream()
-                .map(entry -> {
-                    long productId = entry.getKey();
-                    int quantity = entry.getValue();
-                    Product product = productRepository.findById(productId);
-                    return new Order.ProductQuantity(product, quantity);
-                })
-                .toList();
+            //상품 정보(재고 포함)
+            Product product = productRepository.findById(productId);
+            //주문 수량
+            int orderQuantity = orderProduct.getQuantity();
+
+            return new Order.ProductQuantity(product, orderQuantity);
+        }).toList();
     }
 
-    // ProductService에 재고 확인 로직 추가
+    // 재고 확인 로직 - 메서드명 변경
     public void validateStockAvailability(List<Order.ProductQuantity> productQuantities) {
         for (Order.ProductQuantity pq : productQuantities) {
             Product product = pq.product();
             int quantity = pq.quantity();
-            product.isStockAvailable(quantity);
+            product.checkStockAvailability(quantity);
         }
     }
 
+    @Transactional
     public void deductStock(List<Order.ProductQuantity> productQuantities) {
         for (Order.ProductQuantity pq : productQuantities) {
+            // 상품 정보(재고 포함)
             Product product = pq.product();
+            //현재 재고 정보
+            Inventory inventory = pq.product().getInventory();
+
+            // 주문 수량
             int quantity = pq.quantity();
-            product.deduct(quantity);
-            productRepository.save(product);
+
+            product.deductStock(quantity);
+            // 재고 저장
+            inventoryRepository.save(product.getInventory());
         }
     }
 
     public List<TopSellingProduct> getPopularProducts() {
         return topSellingProductRepository.findAll();
+    }
+
+    public Product save(Product product) {
+        return productRepository.save(product);
     }
 }
