--- src/main/java/kr/hhplus/be/server/interfaces/api/coupon/dto/CouponIssueRequest.java
@@ -1,4 +1,0 @@
-package kr.hhplus.be.server.interfaces.api.coupon.dto;
-
-public record CouponIssueRequest(Long userId, Long couponId) {}
-
