--- src/main/java/kr/hhplus/be/server/interfaces/api/order/dto/OrderItemRequest.java
@@ -1,25 +1,0 @@
-package kr.hhplus.be.server.interfaces.api.order.dto;
-
-import lombok.Getter;
-
-@Getter
-public class OrderItemRequest {
-    private long productId;
-    private int quantity;
-
-    public OrderItemRequest() {}
-
-    public OrderItemRequest(long productId, int quantity) {
-        this.productId = productId;
-        this.quantity = quantity;
-    }
-
-    public long getProductId() {
-        return productId;
-    }
-
-    public int getQuantity() {
-        return quantity;
-    }
-
-}
