--- src/main/java/kr/hhplus/be/server/interfaces/api/point/dto/PointResponse.java
@@ -1,25 +1,0 @@
-package kr.hhplus.be.server.interfaces.api.point.dto;
-
-import kr.hhplus.be.server.domain.point.Point;
-
-import java.math.BigDecimal;
-
-// BalanceResponse
-public class PointResponse {
-
-    private long userId;
-    private long balance;
-
-    public PointResponse(long userId, long amount) {
-        this.userId = userId;
-        this.balance = amount;
-    }
-
-    public long getUserId() {
-        return userId;
-    }
-
-    public long getBalance() {
-        return balance;
-    }
-}
