+++ src/main/java/kr/hhplus/be/server/interfaces/coupon/dto/CouponIssueRequest.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.interfaces.coupon.dto;
+
+public record CouponIssueRequest(Long userId, Long couponId) {}
+
