+++ src/main/java/kr/hhplus/be/server/interfaces/payment/dto/PaymentRequest.java
@@ -1,0 +1,13 @@
+package kr.hhplus.be.server.interfaces.payment.dto;
+
+import lombok.Getter;
+
+@Getter
+public class PaymentRequest {
+    private long orderId;
+
+    public long getOrderId() {
+        return orderId;
+    }
+
+}
